# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "FX Scalper — ICT × Jev × Deriv",
  "summary": "Transcripts → Jev-tagged ICT concepts → typed strategy graph → Deriv backtests and demo execution.",
  "version": "0.2.0"
 },
 "paths": {
...
  "/api/secrets/{name}/test": {
   "post": {
    "tags": [
     "system"
    ],
    "summary": "Test the connection that uses this key",
    "operationId": "test_secret_api_secrets__name__test_post",
    "parameters": [
     {
      "name": "name",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Name"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Test Secret Api Secrets  Name  Test Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/strategy/tagging/jobs": {
   "post": {
    "tags": [
     "strategy"
    ],
    "summary": "Tag transcript windows with Jev (background; progress on /api/stream)",
    "operationId": "start_tagging_api_strategy_tagging_jobs_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TagRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Start Tagging Api Strategy Tagging Jobs Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/strategy/tags": {
   "get": {
    "tags": [
     "strategy"
    ],
    "summary": "Tagged windows ranked by rule-likeness (optionally for one concept)",
    "operationId": "tags_api_strategy_tags_get",
    "parameters": [
     {
      "name": "concept",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Concept"
      }
     },
     {
      "name": "min_rule",
      "in": "query",
      "required": false,
      "schema": {
       "type": "number",
       "default": 0.6,
       "title": "Min Rule"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 60,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Tags Api Strategy Tags Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/strategy/digest": {
   "post": {
    "tags": [
     "strategy"
    ],
    "summary": "Rebuild digest_by_concept.md from stored tags",
    "operationId": "digest_api_strategy_digest_post",
    "parameters": [
     {
      "name": "per_concept",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 25,
       "title": "Per Concept"
      }
     },
     {
      "name": "min_rule",
      "in": "query",
      "required": false,
      "schema": {
       "type": "number",
       "default": 0.6,
       "title": "Min Rule"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Digest Api Strategy Digest Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/strategy/schemas/{strategy_id}": {
   "get": {
    "tags": [
     "strategy"
    ],
    "summary": "One schema (latest version unless ?version=)",
    "operationId": "schema_api_strategy_schemas__strategy_id__get",
    "parameters": [
     {
      "name": "strategy_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Strategy Id"
      }
     },
     {
      "name": "version",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Version"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Schema Api Strategy Schemas  Strategy Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/strategy/schemas/{strategy_id}/graph": {
   "get": {
    "tags": [
     "strategy"
    ],
    "summary": "Decision graph (nodes/edges) for the UI's SVG renderer",
    "operationId": "schema_graph_api_strategy_schemas__strategy_id__graph_get",
    "parameters": [
     {
      "name": "strategy_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Strategy Id"
      }
     },
     {
      "name": "version",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Version"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Schema Graph Api Strategy Schemas  Strategy Id  Graph Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/strategy/schemas/{strategy_id}/mermaid": {
   "get": {
    "tags": [
     "strategy"
    ],
    "summary": "Mermaid flowchart source of the decision graph",
    "operationId": "schema_mermaid_api_strategy_schemas__strategy_id__mermaid_get",
    "parameters": [
     {
      "name": "strategy_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Strategy Id"
      }
     },
     {
      "name": "version",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Version"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "text/plain": {
        "schema": {
         "type": "string"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/backtests/run": {
   "post": {
    "tags": [
     "backtests"
    ],
    "summary": "Start a backtest (background job). An empty body runs the legacy 3-scenario check.",
    "operationId": "run_api_backtests_run_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "anyOf": [
         {
          "additionalProperties": true,
          "type": "object"
         },
         {
          "type": "null"
         }
        ],
        "title": "Body"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/backtests": {
   "get": {
    "tags": [
     "backtests"
    ],
    "summary": "Stored backtest runs (newest first)",
    "operationId": "list_runs_api_backtests_get",
    "parameters": [
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "default": 50,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Runs Api Backtests Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/backtests/{run_id}": {
   "get": {
    "tags": [
     "backtests"
    ],
    "summary": "Full result: summary, validation, breakdowns, node pass rates, trades, equity",
    "operationId": "get_run_api_backtests__run_id__get",
    "parameters": [
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Run Id"
      }
     },
     {
      "name": "include_trades",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": true,
       "title": "Include Trades"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Get Run Api Backtests  Run Id  Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "TagRequest": {
    "properties": {
     "video_ids": {
      "anyOf": [
       {
        "items": {
         "type": "string"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Video Ids",
      "description": "Default: every stored transcript"
     },
     "limit": {
      "anyOf": [
       {
        "type": "integer",
        "minimum": 1.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Limit"
     },
     "retag": {
      "type": "boolean",
      "title": "Retag",
      "default": false
     }
    },
    "type": "object",
    "title": "TagRequest"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 },
...
}